moved logo into separate component
